render-markdown-django-3-0-django-markdown-view/requirements.txt
Django==3.0.8
django-markdown-view==0.0.2
